# Add precompilation workload