This speeds up time-to-first-plot tremendously.

Benchmark (to run in a fresh Julia session):

```julia
using LinearAlgebra, Random, SparseArrays
using SparseMatrixColorings

# simple task
@time begin
    A = sprand(MersenneTwister(0), Bool, 10, 10, 0.2)
    problem = ColoringProblem(; structure=:nonsymmetric, partition=:column)
    algo = GreedyColoringAlgorithm(; decompression=:direct)
    result = coloring(A, problem, algo)
    B = compress(A, result)
    decompress(B, result)
end;

# complicated task
@time begin
    A = sprand(MersenneTwister(0), Bool, 10, 10, 0.2)
    problem = ColoringProblem(; structure=:nonsymmetric, partition=:bidirectional)
    algo = GreedyColoringAlgorithm(; decompression=:substitution, postprocessing=true)
    result = coloring(A, problem, algo)
    Br, Bc = compress(A, result)
    decompress(Br, Bc, result)
end;
```

With precompilation:

- Simple task: `0.000096 seconds (108 allocations: 26.328 KiB)`
- Complicated task: `0.003747 seconds (1.81 k allocations: 112.008 KiB, 98.13% compilation time)`

Without precompilation:

- Simple task: `0.394682 seconds (2.78 M allocations: 141.290 MiB, 1.76% gc time, 99.95% compilation time)`
- Complicated task: `0.896819 seconds (3.00 M allocations: 154.698 MiB, 1.33% gc time, 99.98% compilation time)`